pyproject.toml
[project]
name = "lox"
version = "0.1.0"
description = "Interpretador Lox"
readme = "README.md"
authors = [
    { name = "Fábio Macêdo Mendes", email = "fabiomacedomendes@gmail.com" },
]
requires-python = ">=3.10"
dependencies = ["lark-parser>=0.12.0", "rich>=14.0.0"]

[project.scripts]
lox = "lox.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = ["pytest>=8.3.5"]
